Refactored test helper code to test fixture base class.
Added unit tests for DocumentDbToggle implementations
