+++ test/FeatureToggle.Azure.DocumentDB.Test/DocumentDbToggleTest.cs
@@ -1,0 +1,40 @@
+﻿using FeatureToggle.Azure.Providers;
+using NUnit.Framework;
+using Shouldly;
+using System;
+using System.Collections.Generic;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace FeatureToggle.Azure.DocumentDB.Test
+{
+    [TestFixture]
+    public class DocumentDbToggleTest : DocumentDbTestFixture
+    {
+        [Test]
+        public void FeatureEnabled_ToggleExists_ToggleValueIsFalse()
+        {
+            // Arrange
+            AutoCreateToggle();
+            var toggle = new TestFeatureToggle();
+            // Act
+            var toggleValue = toggle.FeatureEnabled;
+            // Assert
+            toggleValue.ShouldBeFalse();
+        }
+
+        [Test]
+        public async Task FeatureEnabled_ToggleExists_ToggleValueIsTrue()
+        {
+            // Arrange
+            AutoCreateToggle();            
+            await UpdateToggleDocument(new BooleanFeatureToggleDocument(nameof(TestFeatureToggle)) { Enabled = true });
+
+            var toggle = new TestFeatureToggle();
+            // Act
+            var toggleValue = toggle.FeatureEnabled;
+            // Assert
+            toggleValue.ShouldBeTrue();
+        }
+    }
+}
